+++ polars/polars-core/Cargo.toml
@@ -19,6 +19,7 @@ default = ["docs", "temporal", "private"]
 lazy = ["sort_multiple"]
 
 # ~40% faster collect, needed until trustedlength iter stabilizes
+# more fast paths
 performant = []
 
 # extra utilities for Utf8Chunked
