+++ polars/polars-core/src/chunked_array/ops/sort/argsort_multiple.rs
@@ -1,4 +1,5 @@
 use super::*;
+use polars_arrow::data_types::IsFloat;
 
 pub(crate) fn args_validate<T: PolarsDataType>(
     ca: &ChunkedArray<T>,
@@ -23,24 +24,43 @@ pub(crate) fn args_validate<T: PolarsDataType>(
     Ok(())
 }
 
-pub(crate) fn argsort_multiple_impl<T: PartialOrd>(
-    mut vals: Vec<(IdxSize, Option<T>)>,
+fn sort_cmp<T: PartialOrd + IsFloat + Copy>(a: &T, b: &T) -> Ordering {
+    if T::is_float() {
+        match (a.is_nan(), b.is_nan()) {
+            // safety: we checked nans
+            (false, false) => unsafe { a.partial_cmp(b).unwrap_unchecked() },
+            (true, true) => Ordering::Equal,
+            (true, false) => Ordering::Greater,
+            (false, true) => Ordering::Less,
+        }
+    } else {
+        // no floats, so we can compare unchecked
+        unsafe { a.partial_cmp(b).unwrap_unchecked() }
+    }
+}
+
+pub(crate) fn argsort_multiple_impl<T: PartialOrd + Send + IsFloat + Copy>(
+    mut vals: Vec<(IdxSize, T)>,
     other: &[Series],
     reverse: &[bool],
 ) -> Result<IdxCa> {
+    assert_eq!(reverse.len() - 1, other.len());
     let compare_inner: Vec<_> = other
         .iter()
         .map(|s| s.into_partial_ord_inner())
         .collect_trusted();
 
-    vals.sort_by(
-        |tpl_a, tpl_b| match (reverse[0], sort_with_nulls(&tpl_a.1, &tpl_b.1)) {
+    let first_reverse = reverse[0];
+    vals.par_sort_by(
+        |tpl_a, tpl_b| match (first_reverse, sort_cmp(&tpl_a.1, &tpl_b.1)) {
             // if ordering is equal, we check the other arrays until we find a non-equal ordering
             // if we have exhausted all arrays, we keep the equal ordering.
             (_, Ordering::Equal) => {
                 let idx_a = tpl_a.0 as usize;
                 let idx_b = tpl_b.0 as usize;
-                ordering_other_columns(&compare_inner, &reverse[1..], idx_a, idx_b)
+                unsafe {
+                    ordering_other_columns(&compare_inner, reverse.get_unchecked(1..), idx_a, idx_b)
+                }
             }
             (true, Ordering::Less) => Ordering::Greater,
             (true, Ordering::Greater) => Ordering::Less,
