+++ polars/polars-core/src/chunked_array/ops/sort/categorical.rs
@@ -1,6 +1,16 @@
 use super::*;
 use crate::utils::NoNull;
 
+/// Sort with null values, to reverse, swap the arguments.
+fn sort_with_nulls<T: PartialOrd>(a: &Option<T>, b: &Option<T>) -> Ordering {
+    match (a, b) {
+        (Some(a), Some(b)) => a.partial_cmp(b).unwrap(),
+        (None, Some(_)) => Ordering::Less,
+        (Some(_), None) => Ordering::Greater,
+        (None, None) => Ordering::Equal,
+    }
+}
+
 /// Default sorting nulls
 pub fn order_default_null<T: PartialOrd>(a: &Option<T>, b: &Option<T>) -> Ordering {
     sort_with_nulls(a, b)
